+++ src/client/components/teScriber/teScriber.js
@@ -12,8 +12,6 @@ import { pwixI18n } from 'meteor/pwix:i18n';
 
 import { v4 as uuidv4 } from 'uuid';
 
-import 'jquery-resizable-dom/dist/jquery-resizable.min.js';
-
 import 'trumbowyg/dist/ui/trumbowyg.min.css';
 //import 'trumbowyg/dist/trumbowyg.min.js';
 import 'trumbowyg/dist/trumbowyg.js'; // keep this modified version until https://github.com/Alex-D/Trumbowyg/issues/1396 is fixed