+++ src/server/js/check_npms.js
@@ -8,7 +8,6 @@ if( false ){
     // whitelist packages which are included via a subfolder
     require( '@popperjs/core/package.json' );
     require( 'bootstrap/package.json' );
-    require( 'jquery-resizable-dom/package.json' );
     require( 'trumbowyg/package.json' );
     require( 'uuid/package.json' );
 }
@@ -16,8 +15,9 @@ if( false ){
 checkNpmVersions({
     '@popperjs/core': '^2.11.6',
     'bootstrap': '^5.2.1',
-    'jquery-resizable-dom': '^0.35.0',
     'lodash': '^4.17.0',
     'trumbowyg': '^2.25.2',
     'uuid': '^9.0.0'
-    }, 'pwix:editor' );
+},
+    'pwix:editor'
+);